requirements.txt
selenium==3.141.0
pytest==6.2.2
Pillow==7.2.0
allure-pytest==2.8.35
